GameMemory.cs renamed to ProcessMemory.cs
