+++ TombRaider1996/TR1.csproj
@@ -55,7 +55,7 @@
   <ItemGroup>
     <Compile Include="EmulatorData.cs" />
     <Compile Include="GameData.cs" />
-    <Compile Include="GameMemory.cs" />
+    <Compile Include="ProcessMemory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ComponentSettings.cs">
       <SubType>UserControl</SubType>
